+++ CHANGES.rst
@@ -10,7 +10,11 @@ Breaking changes:
 
 New features:
 
-- *add item here*
+- Use get_installer instead of portal_quickinstaller when available, for
+  Plone 5.1 and higher.  [maurits]
+
+- In PloneSandboxLayer make profile upgrade versions persistent.  This
+  way installed profile versions get reset in teardown.  [maurits]
 
 Bug fixes: